requirements (3).txt
```
flask==2.3.3
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
openpyxl==3.1.2
scikit-learn==1.5.0
python-dotenv==1.0.1
```